+++ src/core/library/macroLibraryId.ts
@@ -1,9 +1,10 @@
 import * as vscode from 'vscode';
-import { trimTrailingSep } from './utils';
 
 export type MacroLibraryId = string & { __macroLibraryIdBrand: void };
 
 export function getMacroLibraryId(uri: vscode.Uri): MacroLibraryId {
-  const normalizedUri = uri.with({ fragment: '', query: '', path: trimTrailingSep(uri.path) });
+  const normalizedPath =
+    uri.path !== '/' && uri.path.endsWith('/') ? uri.path.slice(0, -1) : uri.path;
+  const normalizedUri = uri.with({ fragment: '', query: '', path: normalizedPath });
   return normalizedUri.toString(true) as MacroLibraryId;
 }